pyproject.toml
[tool.poetry]
name = "LangGraph_Project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
python-dotenv = "^1.0.1"
langchain = "^0.2.0"
langchain-openai = "^0.1.7"
black = "^24.4.2"
isort = "^5.13.2"
langgraph = "^0.0.53"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"
ipykernel = "^6.29.4"


[tool.poetry.group.D.dependencies]
ipynb = "^0.5.1"
black = {extras = ["jupyter"], version = "^24.4.2"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
